requirements.txt
django = 2.1
djangorestframework = 3.9.2
markdown = 3.0.1
django-filter = 2.1.0
stripe = 2.21.0

